lab3/tests/MapDirectionTest.java
import org.junit.Test;
import org.junit.Assert;
import agh.cs.lab3.MapDirection;

public class MapDirectionTest{
    @Test
    public void nextTest(){
        Assert.assertEquals(MapDirection.EAST, MapDirection.NORTH.next());
        Assert.assertEquals(MapDirection.SOUTH, MapDirection.EAST.next());
        Assert.assertEquals(MapDirection.WEST, MapDirection.SOUTH.next());
        Assert.assertEquals(MapDirection.NORTH, MapDirection.WEST.next());
    }

    @Test
    public void previousTest(){
        Assert.assertEquals(MapDirection.EAST, MapDirection.SOUTH.previous());
        Assert.assertEquals(MapDirection.SOUTH, MapDirection.WEST.previous());
        Assert.assertEquals(MapDirection.WEST, MapDirection.NORTH.previous());
        Assert.assertEquals(MapDirection.NORTH, MapDirection.EAST.previous());
    }
}